# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"MindMate API","version":"0.1.0"},"paths":{"/api/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["auth"],"summary":"Read Users Me","operationId":"read_users_me_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/admin/questions":{"get":{"tags":["admin"],"summary":"Read Questions","operationId":"read_questions_api_admin_questions_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Question"},"title":"Response Read Questions Api Admin Questions Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["admin"],"summary":"Create Question","operationId":"create_question_api_admin_questions_post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Question"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/questions/{question_id}":{"put":{"tags":["admin"],"summary":"Update Question","operationId":"update_question_api_admin_questions__question_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"integer","title":"Question Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Question"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["admin"],"summary":"Delete Question","operationId":"delete_question_api_admin_questions__question_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"question_id","in":"path","required":true,"schema":{"type":"integer","title":"Question Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/resources":{"get":{"tags":["admin"],"summary":"Read Resources","operationId":"read_resources_api_admin_resources_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Resource"},"title":"Response Read Resources Api Admin Resources Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["admin"],"summary":"Create Resource","operationId":"create_resource_api_admin_resources_post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Resource"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/admin/resources/{resource_id}":{"put":{"tags":["admin"],"summary":"Update Resource","operationId":"update_resource_api_admin_resources__resource_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"resource_id","in":"path","required":true,"schema":{"type":"integer","title":"Resource Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Resource"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["admin"],"summary":"Delete Resource","operationId":"delete_resource_api_admin_resources__resource_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"resource_id","in":"path","required":true,"schema":{"type":"integer","title":"Resource Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/profile":{"get":{"tags":["student"],"summary":"Read Profile","operationId":"read_profile_api_profile_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentProfile"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"tags":["student"],"summary":"Update Profile","operationId":"update_profile_api_profile_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentProfileCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentProfile"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/profile/upload-photo":{"post":{"tags":["student"],"summary":"Upload Profile Photo","operationId":"upload_profile_photo_api_profile_upload_photo_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_profile_photo_api_profile_upload_photo_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/questions":{"get":{"tags":["student"],"summary":"Get Active Questions","operationId":"get_active_questions_api_questions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Question"},"type":"array","title":"Response Get Active Questions Api Questions Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/resources":{"get":{"tags":["student"],"summary":"Search Resources","operationId":"search_resources_api_resources_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Q"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Resource"},"title":"Response Search Resources Api Resources Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/assessments/history":{"get":{"tags":["student"],"summary":"Get History","operationId":"get_history_api_assessments_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/recommendations/latest":{"get":{"tags":["student"],"summary":"Get Latest Recommendations","operationId":"get_latest_recommendations_api_recommendations_latest_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/RecommendationResponse"},"type":"array","title":"Response Get Latest Recommendations Api Recommendations Latest Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/assessments":{"post":{"tags":["student"],"summary":"Submit Assessment","operationId":"submit_assessment_api_assessments_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentSubmit"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessmentResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/recommendations/{rec_id}/feedback":{"post":{"tags":["student"],"summary":"Submit Feedback","operationId":"submit_feedback_api_recommendations__rec_id__feedback_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"rec_id","in":"path","required":true,"schema":{"type":"integer","title":"Rec Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendationFeedbackSubmit"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/journal":{"get":{"tags":["student"],"summary":"Get Journal Entries","operationId":"get_journal_entries_api_journal_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["student"],"summary":"Create Journal Entry","operationId":"create_journal_entry_api_journal_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Data"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/journal/{entry_id}":{"put":{"tags":["student"],"summary":"Update Journal Entry","operationId":"update_journal_entry_api_journal__entry_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"entry_id","in":"path","required":true,"schema":{"type":"integer","title":"Entry Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Data"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["student"],"summary":"Delete Journal Entry","operationId":"delete_journal_entry_api_journal__entry_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"entry_id","in":"path","required":true,"schema":{"type":"integer","title":"Entry Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chat":{"post":{"tags":["chat"],"summary":"Chat With Groq","operationId":"chat_with_groq_api_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/debug-env":{"get":{"summary":"Debug Env","operationId":"debug_env_api_debug_env_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/debug-routes":{"get":{"summary":"Debug Routes","operationId":"debug_routes_api_debug_routes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AnswerSubmit":{"properties":{"question_id":{"type":"integer","title":"Question Id"},"answer_value":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Answer Value"},"option_values":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Option Values"}},"type":"object","required":["question_id"],"title":"AnswerSubmit"},"AssessmentResult":{"properties":{"id":{"type":"integer","title":"Id"},"predicted_level":{"type":"string","title":"Predicted Level"},"submitted_at":{"type":"string","format":"date-time","title":"Submitted At"}},"type":"object","required":["id","predicted_level","submitted_at"],"title":"AssessmentResult"},"AssessmentSubmit":{"properties":{"answers":{"items":{"$ref":"#/components/schemas/AnswerSubmit"},"type":"array","title":"Answers"}},"type":"object","required":["answers"],"title":"AssessmentSubmit"},"Body_login_api_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_auth_login_post"},"Body_upload_profile_photo_api_profile_upload_photo_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_profile_photo_api_profile_upload_photo_post"},"ChatMessage":{"properties":{"role":{"type":"string","title":"Role"},"parts":{"items":{"type":"string"},"type":"array","title":"Parts"}},"type":"object","required":["role","parts"],"title":"ChatMessage"},"ChatRequest":{"properties":{"history":{"items":{"$ref":"#/components/schemas/ChatMessage"},"type":"array","title":"History"},"message":{"type":"string","title":"Message"}},"type":"object","required":["history","message"],"title":"ChatRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Question":{"properties":{"text":{"type":"string","title":"Text"},"type":{"type":"string","title":"Type"},"options_json":{"type":"string","title":"Options Json"},"active":{"type":"boolean","title":"Active","default":true},"order_no":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Order No"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["text","type","options_json","id"],"title":"Question"},"QuestionCreate":{"properties":{"text":{"type":"string","title":"Text"},"type":{"type":"string","title":"Type"},"options_json":{"type":"string","title":"Options Json"},"active":{"type":"boolean","title":"Active","default":true},"order_no":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Order No"}},"type":"object","required":["text","type","options_json"],"title":"QuestionCreate"},"QuestionUpdate":{"properties":{"text":{"type":"string","title":"Text"},"type":{"type":"string","title":"Type"},"options_json":{"type":"string","title":"Options Json"},"active":{"type":"boolean","title":"Active","default":true},"order_no":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Order No"}},"type":"object","required":["text","type","options_json"],"title":"QuestionUpdate"},"RecommendationFeedbackSubmit":{"properties":{"rating":{"type":"string","title":"Rating"}},"type":"object","required":["rating"],"title":"RecommendationFeedbackSubmit"},"RecommendationResponse":{"properties":{"id":{"type":"integer","title":"Id"},"rank":{"type":"integer","title":"Rank"},"score":{"type":"number","title":"Score"},"reason":{"type":"string","title":"Reason"},"resource":{"$ref":"#/components/schemas/Resource"}},"type":"object","required":["id","rank","score","reason","resource"],"title":"RecommendationResponse"},"Resource":{"properties":{"title":{"type":"string","title":"Title"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"tags_json":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tags Json"},"duration_min":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration Min"},"content_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Type"},"content_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Url"},"content_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Text"},"thumbnail_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thumbnail Url"},"active":{"type":"boolean","title":"Active","default":true},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["title","id","created_at","updated_at"],"title":"Resource"},"ResourceCreate":{"properties":{"title":{"type":"string","title":"Title"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"tags_json":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tags Json"},"duration_min":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration Min"},"content_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Type"},"content_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Url"},"content_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Text"},"thumbnail_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thumbnail Url"},"active":{"type":"boolean","title":"Active","default":true}},"type":"object","required":["title"],"title":"ResourceCreate"},"ResourceUpdate":{"properties":{"title":{"type":"string","title":"Title"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"tags_json":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tags Json"},"duration_min":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration Min"},"content_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Type"},"content_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Url"},"content_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Text"},"thumbnail_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thumbnail Url"},"active":{"type":"boolean","title":"Active","default":true}},"type":"object","required":["title"],"title":"ResourceUpdate"},"StudentProfile":{"properties":{"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"profile_pic_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Pic Url"},"study_year":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Study Year"},"study_hours":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Study Hours"},"available_time":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Available Time"},"support_preference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Support Preference"},"user_id":{"type":"integer","title":"User Id"}},"type":"object","required":["user_id"],"title":"StudentProfile"},"StudentProfileCreate":{"properties":{"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"profile_pic_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Profile Pic Url"},"study_year":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Study Year"},"study_hours":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Study Hours"},"available_time":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Available Time"},"support_preference":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Support Preference"}},"type":"object","title":"StudentProfileCreate"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"},"user_id":{"type":"integer","title":"User Id"},"role":{"type":"string","title":"Role"}},"type":"object","required":["access_token","token_type","user_id","role"],"title":"Token"},"User":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"role":{"type":"string","title":"Role"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["email","role","id","created_at"],"title":"User"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/login"}}}}}}